+++ samples/Chat/CHANGELOG.md
@@ -1,8 +1,17 @@
 # Change Log - chat
 
-This log was last generated on Fri, 09 Jul 2021 20:41:33 GMT and should not be manually modified.
+This log was last generated on Thu, 22 Jul 2021 17:42:41 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [0.0.1](https://github.com/azure/communication-ui-library/tree/chat_v0.0.1)
+
+Thu, 22 Jul 2021 17:42:41 GMT 
+[Compare changes](https://github.com/azure/communication-ui-library/compare/chat_v0.0.1..chat_v0.0.1)
+
+### Changes
+
+- Bump @azure/communication-react to v1.0.0-beta.3 ([PR #535](https://github.com/azure/communication-ui-library/pull/535) by prprabhu@microsoft.com)
 
 ## [0.0.1](https://github.com/azure/communication-ui-library/tree/chat_v0.0.1)